requirements.txt
python-dotenv
sentence-transformers 
google-genai
pydantic
fastapi 
uvicorn
faiss-cpu 
# faiss-gpu 